Further divide v1 and v2.
